Tighten CI lint coverage: add --all-targets and fmt --check (#27)
Two CI gaps surfaced during v0.3.0 prep:

1. **`cargo clippy --workspace -- -D warnings` lacked `--all-targets`**
— tests, benches, and examples were never lint-checked. Adding it
surfaced ~37 violations across test code in `buffa`, `buffa-types`,
`buffa-codegen`, plus 46 in `buffa-test`.
2. **`cargo fmt --all --check` was never run in CI** — fmt drift from
#20 went undetected (later fixed as a drive-by in #25, so no remaining
drift).

This PR adds both checks to `Taskfile.yml` `lint` and
`.github/workflows/ci.yml` (lint-and-test + msrv-check jobs), then fixes
the violations.

## Fixes by category

| Lint | Count | Resolution |
|---|---|---|
| `field_reassign_with_default` | ~15 in published crates, 35 in
`buffa-test` | Converted to struct literals in published crates;
`#[allow]` on `buffa-test` `mod tests` (the field-assignment style
mirrors the documented protobuf construction pattern) |
| `approx_constant` (`3.14`) | ~12 | Changed to `2.5` in published
crates (arbitrary test data, not PI); allowed in `buffa-test` |
| `items_after_test_module` | 1 | Moved `impl_message.rs` `mod tests` to
end of file (~630 lines of helpers were after it) — zero net lines |
| `identity_op` (`\| 0`, `+ 0`) | 3 | `#[allow]` with comment — the
explicit wire-type / byte-count makes the test encoding readable |
| Assorted (`derivable_impls`, `unnecessary_map_or`, `unnecessary_cast`,
`unnecessary_get_then_check`, `needless_borrows_for_generic_args`,
`clone_on_copy`, `unnecessary_to_owned`) | ~8 | Applied clippy's
suggested fix |

Net +34 lines across 14 files (most of the diff is the `impl_message.rs`
block move). `task lint` and `task test` (1400 tests) both pass.

Aligns with anthropics/connect-rust#25 which made the same
`--all-targets` change there.

+++ buffa/src/types.rs
@@ -1528,8 +1528,7 @@ mod tests {
         let body = borrow_group(&mut cur, 1, crate::RECURSION_LIMIT).unwrap();
 
         // Body should contain the tag for field 2 + the varint 150.
-        let expected_body_len =
-            data.len() - crate::encoding::varint_len(((1u64) << 3) | 4) as usize;
+        let expected_body_len = data.len() - crate::encoding::varint_len(((1u64) << 3) | 4);
         assert_eq!(body.len(), expected_body_len);
         assert!(cur.is_empty());
